feat: API 마이그레이션 및 Pages→Features 리팩터링
- 새 백엔드 API 연동
- 아키텍처 개선 (9개 페이지)
- 코드 정리 및 최적화
#23
