+++ .env
@@ -1,0 +1,1 @@
+REACT_APP_API_BASE_URL=https://eager-emogene-nigonego-9b3dee94.koyeb.app
