+++ src/app/router/Routing.tsx
@@ -1,9 +1,10 @@
-import React from 'react';
+
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import StartSplash from '../../pages/SplashPage/StartSplash';
 import LoginMain from '../../pages/LoginMain/LoginMain';
 import LoginPage from '../../pages/LoginPage/LoginPage';
-import JoinPage from '../../pages/JoinPage/JoinMember';
+import JoinPage from '../../pages/JoinPage/JoinPage';
+import JoinMember from '../../pages/JoinPage/JoinMember';
 import HomeFeed from '../../pages/HomeFeed/HomeFeed';
 import ChatPage from '../../pages/ChatPage/ChatPage';
 import ChatRoom from '../../pages/ChatPage/ChatRoom';
@@ -14,7 +15,6 @@ import YourProfilePage from '../../pages/ProfilePage/YourProfilePage/YourProfile
 import ProfileEditPage from '../../pages/ProfilePage/ProfileEditPage/ProfileEditPage';
 import SearchPage from '../../pages/SearchPage/SearchPage';
 import NotFound from '../../pages/NotFound/NotFound';
-import JoinMember from '../../pages/JoinPage/JoinMember';
 import PostMain from '../../pages/PostPage/PostMain/PostMain';
 import HomeBlank from '../../pages/HomePage/HomePage';
 import MyFollowersPage from '../../pages/ProfilePage/MyProfilePage/MyFollowersPage';
@@ -24,24 +24,24 @@ function Routing() {
     <BrowserRouter>
       <Routes>
         <Route path="/" element={<StartSplash />} />
-        <Route path="/homefeed" element={<HomeFeed />} />
-        <Route path="/home" element={<HomeBlank />} />
         <Route path="/loginmain" element={<LoginMain />} />
         <Route path="/login" element={<LoginPage />} />
         <Route path="/join" element={<JoinPage />} />
         <Route path="/joinmember" element={<JoinMember />} />
-        <Route path="/chat" element={<ChatPage />} />
-        <Route path="/chat/chatroom" element={<ChatRoom />} />
+        <Route path="/homefeed" element={<HomeFeed />} />
         <Route path="/myprofile" element={<MyProfilePage />} />
         <Route path="/profileedit" element={<ProfileEditPage />} />
         <Route path="/myfollowers" element={<MyFollowersPage />} />
         <Route path="/yourprofile" element={<YourProfilePage />} />
         <Route path="/postupload" element={<PostUploadPage />} />
         <Route path="/product" element={<ProductPage />} />
+        <Route path="/postmain" element={<PostMain />} />
+        <Route path="/home" element={<HomeBlank />} />
+        <Route path="/chat" element={<ChatPage />} />
+        <Route path="/chat/chatroom" element={<ChatRoom />} />
         <Route path="/search" element={<SearchPage />} />
         <Route path="/*" element={<NotFound />} />
         {/* <Route path="/test" element={<Product />} /> */}
-        <Route path="/postmain" element={<PostMain />} />
       </Routes>
     </BrowserRouter>
   );
