+++ src/pages/JoinPage/JoinPage.tsx
@@ -1,0 +1,13 @@
+import { Wrapper } from '../LoginPage/LoginPage';
+import JoinContainer from '../../features/auth/components/JoinContainer';
+
+function JoinPage() {
+  return (
+    <Wrapper>
+      <h1>회원가입</h1>
+      <JoinContainer />
+    </Wrapper>
+  );
+}
+
+export default JoinPage;
