+++ src/pages/LoginMain/LoginMain.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+
 import { Link } from 'react-router-dom';
 import {
   Logo,
