+++ src/pages/SplashPage/SplashPage.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+
 import styled from 'styled-components';
 
 function SplashPage() {
