+++ src/shared/api/URL.ts
@@ -1,1 +1,1 @@
-export const URL = 'https://api.mandarin.weniv.co.kr';
+export const URL = process.env.REACT_APP_API_BASE_URL || 'https://eager-emogene-nigonego-9b3dee94.koyeb.app';
