+++ src/shared/lib/axios.tsx
@@ -1,6 +1,6 @@
 import axios from 'axios';
 
-const BASE_URL = 'https://api.mandarin.weniv.co.kr';
+const BASE_URL = process.env.REACT_APP_API_BASE_URL || 'https://eager-emogene-nigonego-9b3dee94.koyeb.app';
 
 const createAxiosInstance = (token?: string) => {
   const headers = token
